fix: allow to call `buildStart` multiple times
From what I gathered, `buildStart` can be called again during the second
build using the same configuration. Since we never cleaned
`buildStartCall`, it resulted in some builds breaking.

This commit tries to address this by wrapping the call to
`resolveFormatter` in try-finally and cleaning `buildStartCall` no
matter the result of the call. Hopefully this works.