Replace Zod with a much simpler normalization and assertion
Zod, despite being a somewhat okay library to perform parsing of inputs
and their validation, is very tolling on the TypeScript performance. It
also bugs TypeScript into emitting huge declarations with erased
documentation.

Unfortunately, the way this project used Zod to perform parsing,
validation and transformation of input module options, has been so
tolling on performance that TypeScript language server couldn't keep up
anymore and started crashing due to heap errors. The suggestions in
editor took seconds to show up. This is a terrible developer experience
that made it significantly harder to maintain the project.

We could've of course looked into the performance bottlenecks, but I
believe it will be only a temporary remedy. TypeScript is just not built
for types of such complexity.

This is why this commit fully replaces Zod with a much simpler approach
to normalization and assertion. Unlike Zod, it's very error prone, and
you can cast some incompatible values to any to cause a crash, however
I believe most developers won't do that and will use TypeScript. Some of
the simple and common assertions are still preserved, so common mistakes
will still be caught.

If you remember, when you had any errors with your options, you would
get Zod validation errors pretty-printed in console. Fear not, this
commit also takes care of that and implements pretty printing of the
assertion errors.

Additional benefit of removing Zod is that all JSDoc documentation is
now fully preserved, which means every single option is documented and
you no longer have to open the documentation site. You also should be
able to import types, normalizers and assertion functions from
`@vintl/nuxt/options` in case need doing so arises.