Export ModuleOptions for Nuxt module builder
Without it, Nuxt module builder will attempt to generate the types
manually and this obviously is not going to work, because the module
project doesn't have a valid Nuxt VIntl configuration, and even if it
did, the type would be inferred from this configuration rather than
using the actual type.

I suppose this was done to fix the issue where they previously inferred
the type based on extends TypeScript comparison on the returned module,
which would fail in some odd cases because the dependency and Nuxt
weren't in sync, so they used different "Module" per se.