Fix TypeScript error about usage of `noEmit`
Composite files are not allowed to use `noEmit`, however, we need to
prevent TypeScript from compiling the files, as this would override JS
files (in our case, eslint.config.mjs). Seems this can be avoided by
emitting declarations only. It's not like were actually going to use
TSC to build the project...